- Update libmseed to v3.4.1. - Update included LICENSE file to Apache v2 to match README.md.

+++ libmseed/test/test-repack.c
@@ -40,7 +40,7 @@ TEST (repack, v3)
 
   CHECK (fd != NULL, "Failed to open output file");
 
-  rv = fwrite (buffer, 1, packedlength, fd);
+  rv = (int)fwrite (buffer, 1, packedlength, fd);
 
   CHECK (rv == packedlength, "Failed to write output file");
   CHECK (fclose (fd) == 0, "Failed to close output file");
@@ -84,7 +84,7 @@ TEST (repack, v2)
 
   CHECK (fd != NULL, "Failed to open output file");
 
-  rv = fwrite (buffer, 1, packedlength, fd);
+  rv = (int)fwrite (buffer, 1, packedlength, fd);
 
   CHECK (rv == packedlength, "Failed to write output file");
   CHECK (fclose (fd) == 0, "Failed to close output file");
